Refactory: numbers_station module
